+++ src/MixedModels.jl
@@ -1,6 +1,7 @@
 module MixedModels
 
 using Arrow
+using BSplineKit
 using DataAPI
 using Distributions
 using GLM
@@ -20,6 +21,7 @@ using StatsBase
 using StatsModels
 using StructTypes
 using Tables
+using TypedTables
 
 using LinearAlgebra: BlasFloat, BlasReal, HermOrSym, PosDefException, copytri!
 using Base: Ryu, require_one_based_indexing
@@ -50,6 +52,7 @@ export @formula,
     LogLink,
     MixedModel,
     MixedModelBootstrap,
+    MixedModelProfile,
     Normal,
     OptSummary,
     Poisson,
@@ -59,6 +62,7 @@ export @formula,
     ReMat,
     SeqDiffCoding,
     SqrtLink,
+    Table,
     UniformBlockDiagonal,
     VarCorr,
     aic,
@@ -72,6 +76,7 @@ export @formula,
     cond,
     condVar,
     condVartables,
+    confint,
     deviance,
     dispersion,
     dispersion_parameter,
@@ -100,9 +105,13 @@ export @formula,
     model_response,
     nobs,
     objective,
+    objective!,
     parametricbootstrap,
     pirls!,
     predict,
+    profile,
+    profileσ,
+    profilevc,
     pwrss,
     ranef,
     raneftables,
@@ -132,6 +141,9 @@ export @formula,
     vcov,
     weights,
     zerocorr
+
+# TODO: move this to the correct spot in list once we've decided on name
+export savereplicates, restorereplicates
 
 """
     MixedModel
@@ -177,7 +189,30 @@ include("blockdescription.jl")
 include("grouping.jl")
 include("mimeshow.jl")
 include("serialization.jl")
+include("profile/profile.jl")
 
-include("precompile_MixedModels.jl")
-_precompile_()
+using PrecompileTools
+
+@setup_workload begin
+    # Putting some things in `setup` can reduce the size of the
+    # precompile file and potentially make loading faster.
+    sleepstudy = MixedModels.dataset(:sleepstudy)
+    contra = MixedModels.dataset(:contra)
+    progress = false
+    @compile_workload begin
+        # all calls in this block will be precompiled, regardless of whether
+        # they belong to your package or not (on Julia 1.8 and higher)
+
+        # these are relatively small models and so shouldn't increase precompile times all that much
+        # while still massively boosting load and TTFX times
+        fit(MixedModel,
+            @formula(reaction ~ 1 + days + (1 + days | subj)),
+            sleepstudy; progress)
+        fit(MixedModel,
+            @formula(use ~ 1 + age + abs2(age) + urban + livch + (1 | urban & dist)),
+            contra,
+            Bernoulli(); progress)
+    end
+end
+
 end # module